+++ packages/mds-compliance-service/tests/index.spec.ts
@@ -116,9 +116,7 @@ describe('ComplianceSnapshots Service Tests', () => {
         end_time: now() - days(3)
       })
     } catch (error: unknown) {
-      expect((error as ServiceErrorDescriptor<'ValidationError'>).details).toContain(
-        'start_time should be less than end_time'
-      )
+      expect((error as ServiceErrorDescriptor).details).toContain('start_time should be less than end_time')
     }
   })