+++ packages/mds-policy-service/tests/policy-client.spec.ts
@@ -27,11 +27,11 @@ describe('spot check unit test policy functions with SimplePolicy', () => {
     })
 
     beforeEach(async () => {
-      await PolicyRepository.deleteAll()
+      await PolicyRepository.truncateAllTables()
     })
 
     afterAll(async () => {
-      await PolicyRepository.deleteAll()
+      await PolicyRepository.truncateAllTables()
       await PolicyRepository.shutdown()
       await GeographyServer.stop()
       await PolicyServer.stop()