+++ packages/mds-provider-service/service/logger.ts
@@ -1,0 +1,3 @@
+import { createLogger } from '@mds-core/mds-logger'
+
+export const ProviderServiceLogger = createLogger('mds-provider-service')