+++ container-images/mds-attachment-service/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @container-images/mds-attachment-service
 
+## 0.2.3
+
+### Patch Changes
+
+- @mds-core/mds-attachment-service@0.5.30
+
+## 0.2.2
+
+### Patch Changes
+
+- @mds-core/mds-attachment-service@0.5.29
+
+## 0.2.1
+
+### Patch Changes
+
+- @mds-core/mds-attachment-service@0.5.28
+
 ## 0.2.0
 
 ### Minor Changes