+++ container-images/mds-collector-api/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @container-images/mds-collector-api
 
+## 0.2.3
+
+### Patch Changes
+
+- @mds-core/mds-collector-api@0.1.76
+
+## 0.2.2
+
+### Patch Changes
+
+- @mds-core/mds-collector-api@0.1.75
+
+## 0.2.1
+
+### Patch Changes
+
+- @mds-core/mds-collector-api@0.1.74
+
 ## 0.2.0
 
 ### Minor Changes