+++ container-images/mds-collector-service/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @container-images/mds-collector-service
 
+## 0.2.3
+
+### Patch Changes
+
+- @mds-core/mds-collector-service@0.4.30
+
+## 0.2.2
+
+### Patch Changes
+
+- @mds-core/mds-collector-service@0.4.29
+
+## 0.2.1
+
+### Patch Changes
+
+- @mds-core/mds-collector-service@0.4.28
+
 ## 0.2.0
 
 ### Minor Changes