+++ container-images/mds-compliance-service/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @container-images/mds-compliance-service
 
+## 0.2.3
+
+### Patch Changes
+
+- @mds-core/mds-compliance-service@0.10.35
+
+## 0.2.2
+
+### Patch Changes
+
+- @mds-core/mds-compliance-service@0.10.34
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [65d1c2b9]
+- Updated dependencies [ffc3c9f7]
+  - @mds-core/mds-compliance-service@0.10.33
+
 ## 0.2.0
 
 ### Minor Changes