+++ container-images/mds-ingest-service/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @container-images/mds-ingest-service
 
+## 0.2.3
+
+### Patch Changes
+
+- @mds-core/mds-ingest-service@0.10.14
+
+## 0.2.2
+
+### Patch Changes
+
+- @mds-core/mds-ingest-service@0.10.13
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [ffc3c9f7]
+  - @mds-core/mds-ingest-service@0.10.12
+
 ## 0.2.0
 
 ### Minor Changes