+++ container-images/mds-policy-service/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @container-images/mds-policy-service
 
+## 0.2.3
+
+### Patch Changes
+
+- @mds-core/mds-policy-service@0.12.3
+
+## 0.2.2
+
+### Patch Changes
+
+- @mds-core/mds-policy-service@0.12.2
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [ffc3c9f7]
+  - @mds-core/mds-policy-service@0.12.1
+
 ## 0.2.0
 
 ### Minor Changes