+++ packages/mds-audit-api/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @mds-core/mds-audit-api
 
+## 0.4.7
+
+### Patch Changes
+
+- @mds-core/mds-attachment-service@0.5.30
+- @mds-core/mds-audit-service@0.6.7
+- @mds-core/mds-ingest-service@0.10.14
+- @mds-core/mds-db@0.6.21
+- @mds-core/mds-api-helpers@0.2.29
+
+## 0.4.6
+
+### Patch Changes
+
+- @mds-core/mds-attachment-service@0.5.29
+- @mds-core/mds-audit-service@0.6.6
+- @mds-core/mds-ingest-service@0.10.13
+- @mds-core/mds-db@0.6.20
+- @mds-core/mds-api-helpers@0.2.28
+
+## 0.4.5
+
+### Patch Changes
+
+- Updated dependencies [65d1c2b9]
+- Updated dependencies [ffc3c9f7]
+  - @mds-core/mds-service-helpers@0.6.10
+  - @mds-core/mds-utils@0.5.2
+  - @mds-core/mds-ingest-service@0.10.12
+  - @mds-core/mds-api-server@0.4.13
+  - @mds-core/mds-attachment-service@0.5.28
+  - @mds-core/mds-audit-service@0.6.5
+  - @mds-core/mds-agency-cache@0.4.25
+  - @mds-core/mds-api-helpers@0.2.27
+  - @mds-core/mds-db@0.6.19
+  - @mds-core/mds-schema-validators@0.5.12
+
 ## 0.4.4
 
 ### Patch Changes