+++ packages/mds-jurisdiction-api/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @mds-core/mds-jurisdiction-api
 
+## 0.1.75
+
+### Patch Changes
+
+- @mds-core/mds-jurisdiction-service@0.5.30
+- @mds-core/mds-api-helpers@0.2.29
+
+## 0.1.74
+
+### Patch Changes
+
+- @mds-core/mds-jurisdiction-service@0.5.29
+- @mds-core/mds-api-helpers@0.2.28
+
+## 0.1.73
+
+### Patch Changes
+
+- Updated dependencies [65d1c2b9]
+- Updated dependencies [e36b8a46]
+  - @mds-core/mds-service-helpers@0.6.10
+  - @mds-core/mds-utils@0.5.2
+  - @mds-core/mds-jurisdiction-service@0.5.28
+  - @mds-core/mds-api-server@0.4.13
+  - @mds-core/mds-api-helpers@0.2.27
+
 ## 0.1.72
 
 ### Patch Changes