+++ packages/mds-policy-author-api/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @mds-core/mds-policy-author-api
 
+## 0.8.3
+
+### Patch Changes
+
+- @mds-core/mds-policy-service@0.12.3
+- @mds-core/mds-api-helpers@0.2.29
+
+## 0.8.2
+
+### Patch Changes
+
+- @mds-core/mds-policy-service@0.12.2
+- @mds-core/mds-api-helpers@0.2.28
+
+## 0.8.1
+
+### Patch Changes
+
+- ffc3c9f7: Create a standard repository method for truncating all table data
+- Updated dependencies [65d1c2b9]
+- Updated dependencies [ffc3c9f7]
+  - @mds-core/mds-service-helpers@0.6.10
+  - @mds-core/mds-utils@0.5.2
+  - @mds-core/mds-policy-service@0.12.1
+  - @mds-core/mds-api-server@0.4.13
+  - @mds-core/mds-api-helpers@0.2.27
+  - @mds-core/mds-schema-validators@0.5.12
+
 ## 0.8.0
 
 ### Minor Changes